Menus Figured Out
Added both a main menu to start the game on, and a page for us to add the control instructions, as well as the coding allowing the player to navigate through all of the menus. Additionally added a script in every scene to both hide the cursor and allow the player to exit the game completely with the press of the escape key. Everything has been tested and works well.