stages/scenes/etc -Pierre
Ok so I added a code to the player scripts so they can only jump on the ground. I also changed their jump height. I added 5 scenes being the 5 stages with platforms all of them are in a platform holder which does nothing except hold the platforms so that way the hierarchy wont look cluttered. I also added a material for the platforms