All Functionality Is Working
The game plays. Everything that is in there works (and as of now I have found no bugs). I had to give up on the randomly switching levels, I could not get some methods to speak with each other, but I was able to set up a new Scene that allows the players to choose a level to start with, and after each score.